Don't add task-list class to lists that have no tasks
Markdown documents can contain multiple lists, some of which might
contain tasks, and some which might not.

Prior to this commit the HTML::Pipeline::TaskList filter added the
'task-list' CSS class to all lists if at least one contained an item
that looked like a task. This made it impossible to style task lists
differently to other lists in the document.